PlanA Ai deploy/requirements.txt
flask
flask-cors
pinecone-client
sentence-transformers
google-genai
gunicorn